Refactor layout to enable easier testing in future
